htz ssh restore + backup

+++ scripts/htz_backup.sh
@@ -4,32 +4,33 @@
 #
 
 DEBUG=false
+MINTER_HAS_ERRORS=false
+
+BACKUP_SSH_HOST=u00000@u00000.your-backup.de
+BACKUP_SSH_PORT=23
+BACKUP_SSH_PRIVKEY_PATH=~/.ssh/id_rsa_storagebox
+BACKUP_MINTER_DATA=${BACKUP_SSH_HOST}:./minter_backups
+
 MINTER_HOME=/home/minter
 MINTER_SERVICE_NAME=minter-node
 MINTER_DATA=${MINTER_HOME}/.minter
-BACKUP_ROOT=${MINTER_HOME}/minter_backups
-MINTER_HAS_ERRORS=false
 
 
 # 0 code - errors not found!
 sudo journalctl -u ${MINTER_SERVICE_NAME}  -n 500 | grep -q panic &&  MINTER_HAS_ERRORS=true
+#
+sudo chmod 0600 ${BACKUP_SSH_PRIVKEY_PATH}
 
 [[ "$DEBUG" == "true" ]] && echo "MINTER_HAS_ERRORS=$MINTER_HAS_ERRORS"
 
 #
-if  [[ "$MINTER_HAS_ERRORS" == "false" ]]; then
+if  [[ "$MINTER_HAS_ERRORS" == "false" ]] && [[ -d  ${MINTER_DATA}/ ]]; then
 
   [[ "$DEBUG" == "true" ]] && echo "=========== Backup started ====="
   #
-  [[ -d  ${BACKUP_ROOT}/ ]] || mkdir -p ${BACKUP_ROOT}/
-  #
-  [[ -d  ${BACKUP_ROOT}/config/ ]] && rm -rf ${BACKUP_ROOT}/config/
-  #
   sudo systemctl stop  ${MINTER_SERVICE_NAME} &&
-  rsync -av --delete --exclude 'config/' ${MINTER_DATA}/ ${BACKUP_ROOT}/ &&
-  rm -rf ${BACKUP_ROOT}/config/ &&
+  rsync -av --delete --progress -e 'ssh -p ${BACKUP_SSH_PORT} -i ${BACKUP_SSH_PRIVKEY_PATH}' --recursive --exclude  'config/' ${MINTER_DATA}/  ${BACKUP_MINTER_DATA}/
   sudo systemctl start  ${MINTER_SERVICE_NAME}
-
   #
   [[ "$DEBUG" == "true" ]] && echo "=========== Backup finished ====="
